Distributed cloud: Fix Patch Orchestration
The dcmanager failed to create patch strategy for sub-clouds.
This problem is caused by keyword argument name mismatch in the
vim interfaces that used by the dcmanager.
The argument names that have compute reference have been
changed to worker on the vim side while the dcmanager
still uses the old names.

This update changes all compute references to worker in
dcmanagerclient.

Closes-Bug:1821081
Depends-On: https://review.openstack.org/#/c/645011/

Change-Id: I04488623ef601422352f77372259e4a514c11681
Signed-off-by: Tao Liu <tao.liu@windriver.com>