Adding upgrade strategy commands
dcmanager strategy-config
dcmanager upgrade-strategy

This change refactors the existing sw-update code so that
patch and upgrades share common functionality.

This adds the strategy 'type' to the CLI output so that a
user can differentiate between which strategy type is
being orchestrated.

patch-strategy-config is generic to all update types, so a new
strategy-config CLI command is now registered.

Change-Id: Icd33eb26f907e8e250ebddbba7d2cebea3592ac7
Depends-On: https://review.opendev.org/#/c/721620
Story: 2007403
Task: 39654
Signed-off-by: albailey <Al.Bailey@windriver.com>