Fix zuul jobs broken due to pip upversion
The bandit defined in base level test-requirements.txt
needed to be restricted to python3.

The version of hacking was conflicting.
Updated pylint to be python3.
The newer hacking pulls in some additional flake8 warnings
which are being suppressed for now.

The gettextutils has per line suppressions based
on the new version of pylint, however that file should be
removed in favor of a solution based on oslo.i18n.

The multiple test-requirements.txt files should be
revisited.

Change-Id: I9de0ad5871865fc3baf79f89e3f9f6bedc523e05
Partial-Bug: #1907678
Signed-off-by: albailey <Al.Bailey@windriver.com>