Deprecate cliutils
The cliutils code dates back to when openstack projects
maintained common code and was later replaced by oslo_utils.

Cliutils module was removed in this patch, and the function
env() was moved to shell.py since that was the only reference
to it.

Additional patches will continue to trim from openstack/common.

Story: 2006796
Task: 41897

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I5ab9771db7356a6d45b9c9e06e49c6fd670b663e