Deprecate apiclient
The apiclient, gettextutils, importutils, strutils, and uuidutils
dates back to when openstack projects maintained common code and was later
replaced with oslo_i18n and oslo_utils.

Here we removed the apiclient code because there was no code using it,
as a result the gettext, importutils, strutils, and uuidutils was
removed as well since apiclient was the only place it was being used.

Story: 2006796
Task: 41901

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I2ce5f8035bb0df37189915308e1f576684d63389