Python 2 to Python 3 compatibility DC
The code was adding two content-types, and when executing with
python3 it got an error from the server, cause it was sending
content-type application json but it was supposed to send
a form data, so the fix was just add a verification to ensure that is
not add 2 content-types if it already exists

Development: When I was trying to find things to modify I followed the
approach of build the client, get the tar file, I set up 2 environments
one based on python2 and another python3, I installed the tar client
in both environments and i exported the env vars that the client expect
to get to request the controller, and doing that I could switch between
the two python and indentifying what I should modify.

Test: After all the modification I built an ISO and I installed that
to run some commands and check if my changes got any side effects. After
that followed the procedure to update the remote CLI docker image and
insert the updated client there and I test this new image in the
remote CLI.

Story: 2007106
Task: 42269

Signed-off-by: Rafael Jardim <rafaeljordao.jardim@windriver.com>
Change-Id: Ibe5c719bfecb09d5c2927d29df43ede59738a563