Adding 'force' support and 'to-version' for kube-upgrade
The 'force' allows a subcloud that is in-sync to be upgraded.
The 'to-version' allows a version to be specified, as long as it
is greater or equal to the available version in the subcloud.

These changes add the ability for subclouds to be orchestrated
independently of the version running on the system controller.

Story: 2008972
Task: 44137
Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/820552
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I62a383712eba3a2809df0af356fe1f41b228ee46