Subcloud-backup CLI outputs list or detail
The subcloud-backup restore and create commands now display the
subcloud(s) information in two different ways. When restoring or
creating a single subcloud, with the --subcloud parameter, the
detailed information for this specific subcloud will be shown. When
restoring a group of subclouds, using the --group parameter, a list
with the information of all subclouds in the group will be displayed.

Test Plan:
1. PASS - Validate that the list containing all subclouds in the
group is displayed when the --group parameter is used.
2. PASS - Validate that the detailed information of one subcloud
is displayed when the --subcloud parameter is used.
3. PASS - Validate that the behaviour of other parameters and
functionalities of the command are unchanged.
4. PASS - Repeat the tests above for both create and restore commands

Story: 2010116
Task: 46778

Signed-off-by: Andre Carneiro <Andre.DexheimerCarneiro@windriver.com>
Change-Id: I18f79e29b83a9d04b60687e83f63bfe394df71bd