Fix subcloud sync status being reported as unknown
The sync_status and endpoint_sync_status attributes are missing from
the from_payload() method, causing the subcloud sync status to be
reported as the default value of 'unknown' when executing the 'subcloud
list' or 'subcloud show' CLI commands.

The error_description argument is not a default keyword argument,
potentially causing issues related to #1995873.

Adds support for sync_status and endpoint_sync_status to from_payload()
method; Changes error_description argument to be a default keyword
argument.

Test plan:
PASS: Execute 'dcmanager subcloud list' and verify if it matches the
sync-status reported by the API;
PASS: Execute 'dcmanager subcloud show subcloud1' and verify if it
matches the current sync-status and endpoint-sync-status in the DB;

Closes-Bug: #1996777
Related-Bug: #1995873

Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>
Change-Id: If6ec5ff706f7c5bf65ff624f6a21d988578ff4bc