Remove error_description field from B&R CLI outputs
Stops the error_description field from being returned as part of the
subcloud-backup create/restore command output.

Test plan:
PASS: Execute 'dcmanager subcloud-backup create ...' and verify that
the error_description is not part of the output message;
PASS: Execute 'dcmanager subcloud-backup restore ...' and verify that
the error_description is not part of the output message;
PASS: Execute 'dcmanager subcloud errors <subcloud>' and verify that
the error_description is still returned correctly.

Closes-Bug: #1997023

Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>
Change-Id: Ic70c42d9d107b4f6fb3be8c33c58f0f19b88e5aa