Add the --upload-only option to dcmanager patch-strategy CLI
This commit adds a new option --upload-only to dcmanager patch-strategy
create. If a patch strategy is created with --upload-only option,
patches will only be uploaded to the subclouds when the strategy is
applied. They will not be installed on the subclouds. The strategy
state is set to 'complete' upon successful upload.

This new option allows the system operator to pre-stage the subclouds
with patches and apply them at a more suitable time following the
existing patch orchestration procedure.

Test Plan:
1. PASS - Run dcmanager help patch-strategy create and verify that the
          --upload-only argument is present;
2. PASS - Verify that the CLI output of dcmanager patch-strategy
          create/delete/show contains the 'upload only' field;
3. PASS - Create and apply a patch strategy using the --upload-only
          option and verify that the strategy completes after uploading
          the patches to the subclouds;
4. PASS - Create and apply a patch strategy without the --upload-only
          option while having uploaded the patches before and check
          that the patches are not uploaded again;
5. PASS - Create and apply a patch strategy without the --upload-only
          option without having uploaded the patches before and check
          that the patches are uploaded correctly.

Depends-On: I045256be2195d7cbf4467e22c3aee06adb2e9e5a

Story: 2010584
Task: 47373

Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>
Change-Id: I951e7fad9a6e59fa5629ced3a909b7e073a5b6bd