Add release optionality to subcloud add/reinstall
Add an optional --release parameter to subcloud add and reinstall
commands to enable release optionality in subcloud add and subcloud
reinstall.

Test Plan:
PASS: Verify that the release parameter was successfully passed to
the backend with the "subcloud add" command when it is present.
PASS: Verify that the release parameter was successfully passed to
the backend with the "subcloud reinstall" command when it is present.
PASS: Verify that the release parameter can be absent in
the "subcloud add" command.
PASS: Verify that the release parameter can be absent in
the "subcloud reinstall" command.

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/877845

Story: 2010611
Task: 47669

Signed-off-by: lzhu1 <li.zhu@windriver.com>
Change-Id: Iaa9abd713352a1936c01cf64c34c50ad56e704ed