Add the --upload-only option to dcmanager upgrade_strategy CLI
This commit adds a new option --upload-only to dcmanager
upgrade-strategy create. If this option is used, the releases
are uploaded to the subclouds and the strategy state is set to complete.

Depends-on: https://review.opendev.org/c/starlingx/distcloud/+/894170

Test Plan:
1. PASS - Run dcmanager help upgrade-strategy create and verify
that the --upload-only argument is present.
2. PASS - Create and apply a patch strategy using the --upload-only
option and verify that the strategy completes after uploading
the patches to the subclouds.
3. PASS - Create and apply a patch strategy without using the
--upload-only option and verify that the strategy continues
after uploading the patches to the subclouds.
4. PASS - Verify that the CLI output of dcmanager upgrade-strategy
create/delete/show/abort/apply contains the 'upload only' field.

Story: 2010676
Task: 48746

Change-Id: I3ab70ff7287ae54fa4c407ef0f3c15784bfc19a6
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>