Fix subcommand help and no-auth verification
This commit fixes the -h/--help functionality for subcommands.
Previously, using -h/--help with subcommands returned the help message
for the base dcmanager command. Now, it correctly displays the help
output specific to the subcommand, as if you had used 'dcmanager help
<subcommand>'.

Additionally, this commit improves the no-auth verification for help
and bash completion commands, resolving an issue where the help
command for subcommands required authentication.

Test Plan:
1. PASS: Verify that the following commands still display the help
message for the base dcmanager command:
- dcmanager -h/--help
- dcmanager help
2. PASS: Verify that the following commands correctly display the help
message for the subcommand:
- dcmanager subcloud list -h/--help
- dcmanager help subcloud list
3. PASS: Without sourcing openrc, run the help and bash completion
commands and confirm that they display the help message without
requiring authentication.

Closes-Bug: 2078944

Change-Id: I04dbf996f5c3efa570d4ed4841f526cfb15e02ff
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>