Fix dcmanager complete execution
The implementation [1] to deprecate commands in dcmanager's client
raised an exception when dcmanager help <command> was executed
which happened to break the dcmanager complete command used during
bootstrap.
This fix solves the issue and adds the deprecation message as a
notice in the relevant commands.

[1]: https://review.opendev.org/c/starlingx/distcloud-client/+/917493

Test plan:
1. PASS: Run dcmanager help <command> displays the relevant
deprecation message
2. PASS: Run dcmanager <command> displays the relevant
deprecation message and does not execute
3. PASS: Run dcmanager complete command

Closes-Bug: 2064677

Change-Id: Icbe6e6f5e8145bb786d4b6aebb804416959cbb67
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>