Add prestage --for-sw-deploy/--for-install options
Add --for-sw-deploy and --for-install option to dcmanager prestage
command.

The --for-install option is the default value. In the future we may want
to remove the --for-install switch entirely, as it is superfluous. For
now, we will leave it in - it can either be explicitly used or omitted
entirely.

Test Cases

NOTE: Testing is only complete for the 'dcmanager subcloud prestage'
commands. There has been no testing with any of the prestage
orchestration commands.

PASS:
- Invoke dcmanager prestage with --for-sw-deploy. Ensure
  for_sw_deploy=true is added to the payload.
- Invoke dcmanager prestage with --for-install. Ensure
  for_install=true is added to the payload.
- Invoke dcmanager prestage with neither option. Ensure
  for_install=true is added to the payload.
- Invoke dcmanager prestage with both options. Ensure
  the prestage operation is rejected.

Story: 2010676
Task: 50323
Depends-on: https://review.opendev.org/c/starlingx/distcloud/+/921634

Change-Id: I9185a4382d00efa32146412a66b0c89b725f29ad
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>