Fix handling of empty subcloud parameter in "dcmanager subcloud show"
This commit addresses a bug that occurs when "dcmanager subcloud show"
 is executed with an empty string as the subcloud parameter.
The client formats the request URL as 'f"/subclouds/{subcloud_ref}"',
leading to a call to "/subclouds/" instead of the intended
subcloud-specific URL. This resulted in the API interpreting the request
as a "subcloud list" instead of a "subcloud show".
The fix raises an error when an empty subcloud parameter is provided.

Test Plan:

PASS: Run dcmanager subcloud show command passing "" as subcloud
value. Verify that an error with message "Subcloud cannot be an empty
value" is raised.

PASS: Run dcmanager subcloud error command passing "" as subcloud
value. Verify that an error with message "Subcloud cannot be an empty
value" is raised.

Closes-Bug: 2077422

Change-Id: Ie62e4e23604cceee7beb5dc7c8586404720e5196
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>