Subcloud show detail command backward compatibility
This commit adds the logic to parse the region_name field returned by the API,
in order to be able to display it in the output of the subcloud show and
subcloud show detail commands according to the release to which the client is
trying to communicate.

Test Plan:
PASS - Run dcmanager subcloud show X
     - Run dcmanager subcloud show --detail X
     - Verify that the region_name field is displayed with that format.

PASS - Get a 24.09 client and connect to a 24.09 System Controller.
     - Run dcmanager subcloud show X
     - Run dcmanager subcloud show --detail X
     - Verify that the region_name field is displayed with that format.

PASS - Get a 24.09 client and connect to a 22.12 System Controller.
     - Run dcmanager subcloud show X
     - Verify that the region_name field is not displayed.
     - Run dcmanager subcloud show X --detail
     - Verify that the region_name field is displayed with that format.

PASS - Get a 22.12 client and connect to a 24.09 System Controller.
     - Run dcmanager subcloud show X
     - Verify that the region_name field is not displayed.
     - Run dcmanager subcloud show X --detail
     - Verify that the region_name field is displayed with that format.

Closes-bug: 2083459

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/931160

Change-Id: I906b3d10bae543c2601745ebba408b5715316112
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>