Use "SystemController" when requesting default strategy config
Previously, retrieving the default strategy configuration required
using "RegionOne" as the subcloud_ref. Since the concept of "RegionOne"
will be removed in the future, and the system controller will only have
a UUID region and the "SystemController" region, this commit updates
references from "RegionOne" to "SystemController" when requesting the
default strategy configuration.

Test Plan:
01. PASS - Run 'dcmanager strategy-config update' using different
    region names: "RegionOne", "SystemController", and without
    specifying a region name. Verify that the default options are
    modified accordingly.
02. PASS - Run the previous test but using 'dcmanager strategy-config
    show' instead.

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/945580

Story: 2011312
Task: 51862

Change-Id: I371fd2550b521943da93c3817c35785bde470cd6
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>