Cache auth information to improve performance
In the first execution, the authentication is performed authenticating
the user with their credentials and service endpoint is discovered.
Then the authentication info is stored in the user's session keyring
and any future execution in this session will reuse the cached
information. When the token expires, the authenticating process will
be executed again and the new auth information is stored again.

The user can refresh the cached information with the --refresh-cache
argument.

The user can disable this feature with the --no-cache argument or
setting the environment variable DCCLIENT_NO_CACHE.

The cache reduced the time spent on authentication and service
discovered from ~350 ms to ~10 ms.

The total time reduced by command invocation is around 1 second.

Test Plan:
- PASS: CLI operations are working correctly
- PASS: Cache is stored and retrieved from the keyring
- PASS: Cached info is used after each operation
- PASS: When the token expires, the authenticating process is executed,
        a new token obtained, cached and then used.
- PASS: Cache is not used if the --no-cache is provided
- PASS: Cache is not used if the DCCLIENT_NO_CACHE environment
        variable is configured
- PASS: Cached authenticating info is reload if the --refresh-cache is
        provided
- PASS: Verify remote client
- PASS: Use users other than sysadmin
- PASS: Verify other OpenStack user
- PASS: Source Keystone admin credentials. Source another user keystone
        credentials in the same session, and verify the cache works

Story: 2011149
Task: 52510
Change-Id: I069565f099bacafecb6b8fa94f3a7b4218276952
Signed-off-by: Wallysson Silva <wallysson.silva@windriver.com>