Add --with-prestage option to sw-deploy-strategy
This commit introduces a new `--with-prestage` parameter to the
`sw-deploy-strategy` command. When specified, the sw-deploy
orchestration will execute a prestage if it was not performed
previously, ensuring all necessary steps are completed before
deployment.

To support this flow, the parameters required by the prestage
operation (`--sysadmin-password` and `--force`) are also added
to the sw-deploy-strategy CLI.

Test Plan:
PASS: Run `sw-deploy-strategy` with `--with-prestage` when
      prestage has not been executed and confirm it is applied
      automatically.
PASS: Run `sw-deploy-strategy` with `--with-prestage` when
      prestage was already executed and confirm no duplicate
      prestage runs.
PASS: Verify that `--sysadmin-password` and `--force` options
      are available and functional with `--with-prestage`.

Story: 2011552
Task: 52899

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/961738

Change-Id: I15720d59630d1cc31938608e2a2fe1b1796d8696
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>